services/processor-service-py/pyproject.toml
[tool.poetry]
name = "processor-service"
version = "0.1.0"
description = "Blog post processing service"
authors = ["Your Name <your.email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
google-cloud-pubsub = "^2.19.0"
google-cloud-firestore = "^2.14.0"
google-cloud-storage = "^2.14.0"
openai = "^1.0.0"
python-json-logger = "^2.0.7"
google-genai = "^1.19.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
black = "^23.12.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3 